Move monero/bitcoin modules to files
