+++ swap/src/bitcoin.rs
@@ -1,0 +1,1 @@
+pub use bitcoin::Amount;
